+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/asl/model/AslStructureColumn.java
@@ -18,6 +18,7 @@
 package org.ehrbase.openehr.aqlengine.asl.model;
 
 import static org.ehrbase.jooq.pg.Tables.COMP_VERSION;
+import static org.ehrbase.jooq.pg.Tables.EHR_FOLDER_VERSION;
 
 import java.time.OffsetDateTime;
 import java.util.UUID;
@@ -40,6 +41,9 @@ public enum AslStructureColumn {
     RM_ENTITY(ObjectDataTablePrototype.INSTANCE.RM_ENTITY, String.class, false),
     TEMPLATE_ID(COMP_VERSION.TEMPLATE_ID, UUID.class, AslExtractedColumn.TEMPLATE_ID, true),
     SYS_VERSION(ObjectVersionTablePrototype.INSTANCE.SYS_VERSION, Integer.class, true),
+
+    // Columns for FOLDER querying
+    EHR_FOLDER_IDX(EHR_FOLDER_VERSION.EHR_FOLDERS_IDX, Integer.class, true),
 
     // Columns for VERSION querying
     AUDIT_ID(ObjectVersionTablePrototype.INSTANCE.AUDIT_ID, UUID.class, true),