+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/sql/ConditionUtils.java
@@ -34,7 +34,6 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Stream;
 import javax.annotation.Nonnull;
-import org.apache.commons.lang3.NotImplementedException;
 import org.apache.commons.lang3.tuple.Pair;
 import org.ehrbase.jooq.pg.util.AdditionalSQLFunctions;
 import org.ehrbase.openehr.aqlengine.asl.model.AslRmTypeAndConcept;
@@ -49,6 +48,7 @@ import org.ehrbase.openehr.aqlengine.asl.model.condition.AslNotNullQueryConditio
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslNotQueryCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslOrQueryCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslPathChildCondition;
+import org.ehrbase.openehr.aqlengine.asl.model.condition.AslProvidesJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslQueryCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslQueryCondition.AslConditionOperator;
 import org.ehrbase.openehr.aqlengine.asl.model.condition.AslTrueQueryCondition;
@@ -57,9 +57,11 @@ import org.ehrbase.openehr.aqlengine.asl.model.field.AslColumnField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslComplexExtractedColumnField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslConstantField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslField;
+import org.ehrbase.openehr.aqlengine.asl.model.field.AslFolderItemIdVirtualField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslSubqueryField;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslAuditDetailsJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslDelegatingJoinCondition;
+import org.ehrbase.openehr.aqlengine.asl.model.join.AslFolderItemJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoin;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslPathFilterJoinCondition;
@@ -76,6 +78,17 @@ import org.jooq.impl.DSL;
 
 final class ConditionUtils {
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+
+    private static final EnumSet<AslSourceRelation> SUPPORTED_DESCENDANT_PARENT_RELATIONS = EnumSet.of(
+            AslSourceRelation.COMPOSITION,
+            AslSourceRelation.EHR_STATUS,
+            AslSourceRelation.FOLDER,
+            AslSourceRelation.EHR);
+    private static final EnumSet<AslSourceRelation> SUPPORTED_DESCENDANT_CONDITIONS = EnumSet.of(
+            AslSourceRelation.COMPOSITION,
+            AslSourceRelation.EHR_STATUS,
+            AslSourceRelation.FOLDER // FOLDER CONTAINS FOLDER
+            );
 
     private ConditionUtils() {}
 
@@ -104,6 +117,8 @@ final class ConditionUtils {
                                 AUDIT_DETAILS.ID.getName(),
                                 UUID.class,
                                 true)));
+                case AslFolderItemJoinCondition c -> conditions.add(
+                        joinFolderItemIdEqualVoIdCondition(c, sqlLeft, sqlRight));
             }
         }
 
@@ -113,9 +128,9 @@ final class ConditionUtils {
     private static void addDelegatingJoinConditions(
             AslDelegatingJoinCondition joinCondition, List<Condition> conditions, Table<?> sqlLeft, Table<?> sqlRight) {
         (switch (joinCondition.getDelegate()) {
+                    case AslPathChildCondition c -> pathChildConditions(c, sqlLeft, sqlRight, true);
                     case AslEntityIdxOffsetCondition c -> entityIdxOffsetConditions(c, sqlLeft, sqlRight, true);
                     case AslDescendantCondition c -> descendantConditions(c, sqlLeft, sqlRight, true);
-                    case AslPathChildCondition c -> pathChildConditions(c, sqlLeft, sqlRight, true);
                 })
                 .forEach(conditions::add);
     }
@@ -126,54 +141,29 @@ final class ConditionUtils {
             final Table<?> sqlRight,
             final boolean isJoinCondition) {
         AslSourceRelation parentRelation = dc.getParentRelation();
-        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS)
+        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS, AslSourceRelation.FOLDER)
                 .contains(parentRelation)) {
             throw new IllegalArgumentException("unexpected parent relation type %s".formatted(parentRelation));
         }
-        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS)
+        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS, AslSourceRelation.FOLDER)
                 .contains(dc.getChildRelation())) {
             throw new IllegalArgumentException(
                     "unexpected descendant relation type %s".formatted(dc.getChildRelation()));
         }
 
         return switch (parentRelation) {
-            case COMPOSITION, EHR_STATUS -> {
-                AslStructureColumn pKeyField = parentRelation == AslSourceRelation.COMPOSITION
-                        ? AslStructureColumn.VO_ID
-                        : AslStructureColumn.EHR_ID;
-                yield Stream.of(
-                        // l.pKey == r.pKey
-                        FieldUtils.field(
-                                        sqlLeft,
-                                        dc.getLeftProvider(),
-                                        dc.getLeftOwner(),
-                                        pKeyField.getFieldName(),
-                                        UUID.class,
-                                        true)
-                                .eq(FieldUtils.field(
-                                        sqlRight,
-                                        dc.getRightProvider(),
-                                        dc.getRightOwner(),
-                                        pKeyField.getFieldName(),
-                                        UUID.class,
-                                        isJoinCondition)),
-                        // l.num == r.parent_num
-                        FieldUtils.field(
-                                        sqlLeft,
-                                        dc.getLeftProvider(),
-                                        dc.getLeftOwner(),
-                                        AslStructureColumn.NUM.getFieldName(),
-                                        Integer.class,
-                                        true)
-                                .eq(FieldUtils.field(
-                                        sqlRight,
-                                        dc.getRightProvider(),
-                                        dc.getRightOwner(),
-                                        AslStructureColumn.PARENT_NUM.getFieldName(),
-                                        Integer.class,
-                                        isJoinCondition)));
-            }
-            case FOLDER -> throw new NotImplementedException("Joining FOLDER is not yet supported");
+            case EHR_STATUS -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.EHR_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumEqualParentNumCondition(dc, sqlLeft, sqlRight, isJoinCondition));
+                // l.vo_id == r.vo_id and l.num == r.parent_num
+            case COMPOSITION -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.VO_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumEqualParentNumCondition(dc, sqlLeft, sqlRight, isJoinCondition));
+                // l.ehr_id == r.ehr_id and l.folder_idx = r.folder_idx and l.num == r.parent_num
+            case FOLDER -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.EHR_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinColumnEqualCondition(AslStructureColumn.EHR_FOLDER_IDX, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumEqualParentNumCondition(dc, sqlLeft, sqlRight, isJoinCondition));
             case AUDIT_DETAILS -> throw new IllegalArgumentException(
                     "Path child condition not applicable to AUDIT_DETAILS");
             case EHR -> throw new IllegalArgumentException("Path child condition not applicable to EHR");
@@ -204,14 +194,12 @@ final class ConditionUtils {
 
         // TODO cleanup
         AslSourceRelation parentRelation = dc.getParentRelation();
-        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS, AslSourceRelation.EHR)
-                .contains(parentRelation)) {
+        if (!SUPPORTED_DESCENDANT_PARENT_RELATIONS.contains(parentRelation)) {
             throw new IllegalArgumentException("unexpected parent relation type %s".formatted(parentRelation));
         }
-        if (!EnumSet.of(AslSourceRelation.COMPOSITION, AslSourceRelation.EHR_STATUS)
-                .contains(dc.getDescendantRelation())) {
-            throw new IllegalArgumentException(
-                    "unexpected descendant relation type %s".formatted(dc.getDescendantRelation()));
+        AslSourceRelation descendantRelation = dc.getDescendantRelation();
+        if (!SUPPORTED_DESCENDANT_CONDITIONS.contains(descendantRelation)) {
+            throw new IllegalArgumentException("unexpected descendant relation type %s".formatted(descendantRelation));
         }
 
         return switch (parentRelation) {
@@ -224,52 +212,18 @@ final class ConditionUtils {
                                     AslStructureColumn.EHR_ID.getFieldName(),
                                     UUID.class,
                                     isJoinCondition)));
-            case COMPOSITION, EHR_STATUS -> {
-                AslStructureColumn pKeyField = parentRelation == AslSourceRelation.COMPOSITION
-                        ? AslStructureColumn.VO_ID
-                        : AslStructureColumn.EHR_ID;
-                yield Stream.of(
-                        // l.pKey == r.pKey
-                        FieldUtils.field(
-                                        sqlLeft,
-                                        dc.getLeftProvider(),
-                                        dc.getLeftOwner(),
-                                        pKeyField.getFieldName(),
-                                        UUID.class,
-                                        true)
-                                .eq(FieldUtils.field(
-                                        sqlRight,
-                                        dc.getRightProvider(),
-                                        dc.getRightOwner(),
-                                        pKeyField.getFieldName(),
-                                        UUID.class,
-                                        isJoinCondition)),
-                        // l.num < r.num <= l.num_cap
-                        FieldUtils.field(
-                                        sqlRight,
-                                        dc.getRightProvider(),
-                                        dc.getRightOwner(),
-                                        AslStructureColumn.NUM.getFieldName(),
-                                        Integer.class,
-                                        true)
-                                .between(
-                                        FieldUtils.field(
-                                                        sqlLeft,
-                                                        dc.getLeftProvider(),
-                                                        dc.getLeftOwner(),
-                                                        AslStructureColumn.NUM.getFieldName(),
-                                                        Integer.class,
-                                                        isJoinCondition)
-                                                .add(DSL.inline(1)),
-                                        FieldUtils.field(
-                                                sqlLeft,
-                                                dc.getLeftProvider(),
-                                                dc.getLeftOwner(),
-                                                AslStructureColumn.NUM_CAP.getFieldName(),
-                                                Integer.class,
-                                                isJoinCondition)));
-            }
-            case FOLDER -> throw new NotImplementedException("Joining FOLDER is not yet supported");
+            case EHR_STATUS -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.EHR_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumCapBetweenCondition(dc, sqlLeft, sqlRight, isJoinCondition));
+                // l.vo_id == r.vo_id and l.num < r.num <= l.num_cap
+            case COMPOSITION -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.VO_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumCapBetweenCondition(dc, sqlLeft, sqlRight, isJoinCondition));
+                // l.ehr_id == r.ehr_id and l.folder_idx == r.folder_idx and l.num < r.num <= l.num_cap
+            case FOLDER -> Stream.of(
+                    joinColumnEqualCondition(AslStructureColumn.EHR_ID, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinColumnEqualCondition(AslStructureColumn.EHR_FOLDER_IDX, dc, sqlLeft, sqlRight, isJoinCondition),
+                    joinNumCapBetweenCondition(dc, sqlLeft, sqlRight, isJoinCondition));
             case AUDIT_DETAILS -> throw new IllegalArgumentException(
                     "Descendant condition not applicable to AUDIT_DETAILS");
         };
@@ -372,6 +326,8 @@ final class ConditionUtils {
             case AslAggregatingField __ -> throw new IllegalArgumentException(
                     "AslAggregatingField cannot be used in WHERE");
             case AslSubqueryField __ -> throw new IllegalArgumentException("AslSubqueryField cannot be used in WHERE");
+            case AslFolderItemIdVirtualField __ -> throw new IllegalArgumentException(
+                    "AslFolderItemIdValuesColumnField cannot be used in WHERE");
         };
     }
 
@@ -544,6 +500,95 @@ final class ConditionUtils {
                 default -> throw new IllegalStateException("Unexpected value: " + operator);
             };
         };
+    }
+
+    /**
+     * Provides a join conditions using the given column name
+     * <code>[sqlLeft].column = [sqlRight].column</code>
+     * Example:
+     * <code>"p_data__0"."p_data__0_vo_id" = "p_events__0"."p_events__0_vo_id"</code>
+     */
+    private static Condition joinColumnEqualCondition(
+            AslStructureColumn column,
+            AslProvidesJoinCondition dc,
+            Table<?> sqlLeft,
+            Table<?> sqlRight,
+            boolean aliased) {
+        final String cName = column.getFieldName();
+        return FieldUtils.field(sqlLeft, dc.getLeftProvider(), dc.getLeftOwner(), cName, UUID.class, true)
+                .eq(FieldUtils.field(sqlRight, dc.getRightProvider(), dc.getRightOwner(), cName, UUID.class, aliased));
+    }
+
+    /**
+     * Provides a parent child join conditions using the left <code>num</code> to right <code>parent_num</code>
+     * <code>[sqlLeft].num = [sqlRight].parent_name</code>
+     * Example:
+     * <code>"p_data__0"."p_data__0_vo_id" = "p_events__0"."p_events__0_vo_id"</code>
+     */
+    private static Condition joinNumEqualParentNumCondition(
+            AslProvidesJoinCondition dc, Table<?> sqlLeft, Table<?> sqlRight, boolean aliased) {
+
+        final String num = AslStructureColumn.NUM.getFieldName();
+        final String parentNum = AslStructureColumn.PARENT_NUM.getFieldName();
+
+        return FieldUtils.field(sqlLeft, dc.getLeftProvider(), dc.getLeftOwner(), num, Integer.class, true)
+                .eq(FieldUtils.field(
+                        sqlRight, dc.getRightProvider(), dc.getRightOwner(), parentNum, Integer.class, aliased));
+    }
+
+    /**
+     * Provides a parent child join conditions using the left <code>num</code> to right <code>parent_num</code>
+     * <code>[sqlRight].num between ([sqlLeft].num + 1) and [sqlLeft].num_cap</code>
+     * Example:
+     * <code>"sAN_d_0"."sAN_d_0_num" between ("sCO_c_0"."sCO_c_0_num" + 1) and "sCO_c_0"."sCO_c_0_num_cap"</code>
+     */
+    private static Condition joinNumCapBetweenCondition(
+            AslProvidesJoinCondition dc, Table<?> sqlLeft, Table<?> sqlRight, boolean aliased) {
+
+        final String numFieldName = AslStructureColumn.NUM.getFieldName();
+        final String numCapFieldName = AslStructureColumn.NUM_CAP.getFieldName();
+
+        final AslQuery leftProvider = dc.getLeftProvider();
+        final AslQuery leftOwner = dc.getLeftOwner();
+
+        return FieldUtils.field(sqlRight, dc.getRightProvider(), dc.getRightOwner(), numFieldName, Integer.class, true)
+                .between(
+                        FieldUtils.field(sqlLeft, leftProvider, leftOwner, numFieldName, Integer.class, aliased)
+                                .add(DSL.inline(1)),
+                        FieldUtils.field(sqlLeft, leftProvider, leftOwner, numCapFieldName, Integer.class, aliased));
+    }
+
+    /**
+     * Provides the FOLDER contains COMPOSITION join condition using
+     * <code>[sqlRight]_vo_id = [sqlLeft]_item_id_value</code>
+     * Example:
+     * <code>on "sCO_c_0_vo_id" = "sF_0"."sF_0_item_id_value"</code>
+     *
+     * @param dc {@link AslFolderItemJoinCondition}
+     * @param sqlLeft structure query on <code>folder_data</code>
+     * @param sqlRight structure query on <code>comp_data</code>
+     * @return joinByItemId matching the composition void against the folder item id
+     */
+    private static Condition joinFolderItemIdEqualVoIdCondition(
+            AslFolderItemJoinCondition dc, Table<?> sqlLeft, Table<?> sqlRight) {
+
+        AslQuery leftOwner = dc.getLeftOwner();
+
+        AslQuery rightProvider = dc.rightProvider();
+        AslQuery rightOwner = dc.getRightOwner();
+
+        AslFolderItemIdVirtualField column = leftOwner.getSelect().stream()
+                .filter(AslFolderItemIdVirtualField.class::isInstance)
+                .map(AslFolderItemIdVirtualField.class::cast)
+                .findFirst()
+                .orElseThrow(() -> new IllegalStateException(
+                        "AslFolderItemJoinCondition requires an AslFolderItemIdValuesColumnField"));
+
+        // comp.vo_id == folder.data /items/id/value
+        // on "sCO_c_0_vo_id" = "sF_0_data_item_id_value"
+        return FieldUtils.field(
+                        sqlRight, rightProvider, rightOwner, AslStructureColumn.VO_ID.getFieldName(), UUID.class, true)
+                .eq(FieldUtils.field(sqlLeft, column, column.getFieldName(), UUID.class, true));
     }
 
     static String escapeAsJsonString(String string) {