+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [unreleased]
  ### Added
 * Feature toggle for ehrscape API (ehrbase.rest.ehrscape.enabled, default: false) ([#1415](https://github.com/ehrbase/ehrbase/pull/1415))
+* Add experimental `AQL` support for `FOLDER` ([#1401](https://github.com/ehrbase/ehrbase/pull/1401))
  ### Changed
  ### Fixed